install plugins in background

+++ src/odcommon/usr/local/opendomo/services/syscript/installPlugin.sh
@@ -5,6 +5,7 @@
 # Copyright(c) 2014 OpenDomo Services SL. Licensed under GPL v3 or later
 
 REPOFILE="/var/opendomo/tmp/repo.lst"
+PIDFILE="/var/opendomo/tmp/installplugin.pid"
 
 if ! test -f "$REPOFILE"
 then
@@ -12,12 +13,16 @@ then
 	exit 1
 fi
 
-
 if test -z "$1"
 then
 	echo "#ERROR You must specify the plugin ID"
 	exit 2
 else
+	if test -f "$PIDFILE"
+	then
+		echo "#ERROR System is busy. Try later"
+		exit 3
+	fi
 	echo "#> Install plugin"
 	echo "list:$0"
 	SIZE=`grep ^$1 $REPOFILE | sort | head -n1 | cut -f7 -d';' `
@@ -28,7 +33,7 @@ else
 		exit 3
 	else
 		echo "#INFO Installing plugin $1"
-		/usr/local/bin/plugin_add.sh $1 >/dev/null 2>/dev/null
+		/usr/local/bin/bgshell "/usr/local/bin/plugin_add.sh $1 "
 	fi
 	
 	echo "actions:"
